creates users route and moves endpoints
